+++ crates/selene-core/src/gqlstatus.rs
@@ -26,6 +26,7 @@ pub const ALL_GQLSTATUS_NAMES: &[(&str, &str)] = &[
     ("22G04", "values-not-comparable"),
     ("22G05", "invalid-date-time-or-datetime-function-field-name"),
     ("22G06", "invalid-date-time-or-datetime-function-value"),
+    ("22G07", "invalid-duration-function-field-name"),
     ("22G0C", "list-element-error"),
     ("22G0M", "multiple-assignments-to-graph-element-property"),
     (